refactor: consolidate imports and normalize Chinese comment punctuation
- Use barrel exports instead of deep import paths
- Add shape module and text-cache re-exports in index files
- Normalize Chinese comments to full-width punctuation
- Add deprecation version to Composition.rootItem
- Exclude temp/ from tsconfig.check.json
