+++ packages/effects-core/src/math/index.ts
@@ -6,6 +6,11 @@ export * from './shape/build-line';
 export * from './shape/graphics-path';
 export * from './shape/ellipse';
 export * from './shape/poly-star';
+export * from './shape/circle';
 export * from './shape/polygon';
+export * from './shape/rectangle';
+export * from './shape/triangle';
 export * from './shape/shape-path';
+export * from './shape/shape-primitive';
+export * from './shape/build-adaptive-bezier';
 export * from './bezier';
