+++ packages/effects-core/src/render/graphics.ts
@@ -6,14 +6,8 @@ import type { Engine } from '../engine';
 import { glContext } from '../gl';
 import { Geometry } from './geometry';
 import { Material } from '../material';
-import { buildLine } from '../math';
-import type { StrokeAttributes } from '../math';
-import { buildAdaptiveBezier } from '../math/shape/build-adaptive-bezier';
-import { Circle } from '../math/shape/circle';
-import { Polygon } from '../math/shape/polygon';
-import { Rectangle } from '../math/shape/rectangle';
-import type { ShapePrimitive } from '../math/shape/shape-primitive';
-import { Triangle } from '../math/shape/triangle';
+import type { StrokeAttributes, ShapePrimitive } from '../math';
+import { buildLine, buildAdaptiveBezier, Circle, Polygon, Rectangle, Triangle } from '../math';
 import type { Texture } from '../texture';
 import type { FontStyle, FontWeight } from './text-cache';
 import { ATLAS_SIZE, TextCache } from './text-cache';
@@ -52,7 +46,7 @@ export class Graphics {
     miterLimit: 10,
   };
 
-  // 当前批次状态:批次切换时(类型变化或纹理变化)需要 flush
+  // 当前批次状态：批次切换时(类型变化或纹理变化)需要 flush
   private currentBatchType: BatchType = 'colored';
   private currentBatchTexture: Texture | null = null;
 
@@ -166,7 +160,7 @@ export class Graphics {
   }
 
   /**
-   * 清空缓冲区,准备新的绘制批次
+   * 清空缓冲区，准备新的绘制批次
    */
   begin (): void {
     this.indices.length = 0;
@@ -180,10 +174,10 @@ export class Graphics {
     this.currentBatchType = 'colored';
     this.currentBatchTexture = null;
 
-    // 创建从屏幕坐标到 NDC 的投影矩阵,屏幕坐标: (0, 0) 在左下角, (width, height) 在右上角
+    // 创建从屏幕坐标到 NDC 的投影矩阵，屏幕坐标: (0, 0) 在左下角，(width, height) 在右上角
     const { width, height } = this.engine.canvas.getBoundingClientRect();
 
-    // 正交投影矩阵:将屏幕坐标 [0, width] x [0, height] 映射到 NDC [-1, 1] x [-1, 1]
+    // 正交投影矩阵：将屏幕坐标 [0, width] x [0, height] 映射到 NDC [-1, 1] x [-1, 1]
     const projectionMatrix = new Matrix4(
       2 / width, 0, 0, 0,  // 第一列
       0, 2 / height, 0, 0,  // 第二列
@@ -196,7 +190,7 @@ export class Graphics {
   }
 
   /**
-   * 将当前变换压入栈,并设置新的变换
+   * 将当前变换压入栈，并设置新的变换
    * @param transform - 新的变换矩阵(会与当前变换相乘)
    */
   pushTransform (transform: Matrix3): void {
@@ -227,7 +221,7 @@ export class Graphics {
   }
 
   /**
-   * 切换到指定批次类型/纹理。若与当前批次不一致,先 flush 已累积顶点
+   * 切换到指定批次类型/纹理。若与当前批次不一致，先 flush 已累积顶点
    */
   private ensureBatch (type: BatchType, texture: Texture | null = null): void {
     const sameBatch = this.currentBatchType === type
@@ -242,15 +236,15 @@ export class Graphics {
   }
 
   /**
-   * 提交当前累积的顶点到 renderer,清空缓冲(批次内部 flush 不重置 batchType)
+   * 提交当前累积的顶点到 renderer，清空缓冲(批次内部 flush 不重置 batchType)
    */
   private flushBatch (): void {
     if (this.currentVertexCount === 0 || this.currentIndexCount === 0) {
       return;
     }
 
-    // 把所有 dirty 的字符 atlas 内容上传到 GPU,确保即将采样的 textured 批次拿到最新像素;
-    // 一帧多次 drawText 共写同一 atlas 只在 flush 边界 upload 一次,避免每次 drawText 都重传整张 canvas
+    // 把所有 dirty 的字符 atlas 内容上传到 GPU，确保即将采样的 textured 批次拿到最新像素;
+    // 一帧多次 drawText 共写同一 atlas 只在 flush 边界 upload 一次，避免每次 drawText 都重传整张 canvas
     this.textCache.uploadDirty();
 
     const verticesArray = new Float32Array(this.vertices);
@@ -285,13 +279,13 @@ export class Graphics {
 
   /**
    * 线段顶点按顺序连接 (p0-p1, p1-p2, p2-p3, ...)
-   * @param points - 点数组,格式 [x1,y1,x2,y2,...],至少需要2个点(4个数值)
-   * @param color - 线条颜色,范围 0-1
+   * @param points - 点数组，格式 [x1,y1,x2,y2,...]，至少需要2个点(4个数值)
+   * @param color - 线条颜色，范围 0-1
    * @param thickness - 线宽(像素)
    */
   drawLines (points: number[], color: Color = Color.WHITE, thickness: number = 1.0): void {
     if (!points || points.length < 4 || points.length % 2 !== 0) {
-      console.warn('drawLines: 至少需要2个点(4个数值), 且数组长度必须为偶数');
+      console.warn('drawLines: 至少需要2个点（4个数值），且数组长度必须为偶数');
 
       return;
     }
@@ -317,8 +311,8 @@ export class Graphics {
    * @param y1 - 起点 Y 坐标
    * @param x2 - 终点 X 坐标
    * @param y2 - 终点 Y 坐标
-   * @param color - 线条颜色,默认白色,范围 0-1
-   * @param thickness - 线宽(像素),默认 1
+   * @param color - 线条颜色，默认白色，范围 0-1
+   * @param thickness - 线宽(像素)，默认 1
    */
   drawLine (x1: number, y1: number, x2: number, y2: number, color: Color = Color.WHITE, thickness: number = 1.0): void {
     const linePoints = this.lineShape.points;
@@ -331,7 +325,7 @@ export class Graphics {
   }
 
   /**
-   * 绘制三阶贝塞尔曲线(P0 起点,P1/P2 控制点,P3 终点)
+   * 绘制三阶贝塞尔曲线（P0 起点，P1/P2 控制点，P3 终点）
    * @param x1 - P0 起点 X 坐标
    * @param y1 - P0 起点 Y 坐标
    * @param x2 - P1 第一控制点 X 坐标
@@ -340,8 +334,8 @@ export class Graphics {
    * @param y3 - P2 第二控制点 Y 坐标
    * @param x4 - P3 终点 X 坐标
    * @param y4 - P3 终点 Y 坐标
-   * @param color - 曲线颜色,默认白色,范围 0-1
-   * @param thickness - 线宽(像素),默认 1
+   * @param color - 曲线颜色，默认白色，范围 0-1
+   * @param thickness - 线宽(像素)，默认 1
    */
   drawBezier (x1: number, y1: number, x2: number, y2: number, x3: number, y3: number, x4: number, y4: number, color: Color = Color.WHITE, thickness: number = 1.0): void {
     const bezierPoints = this.bezierShape.points;
@@ -355,15 +349,15 @@ export class Graphics {
   }
 
   /**
-   * 绘制三角形边框(闭合连接 3 个顶点)
+   * 绘制三角形边框（闭合连接 3 个顶点）
    * @param x1 - 顶点 1 X 坐标
    * @param y1 - 顶点 1 Y 坐标
    * @param x2 - 顶点 2 X 坐标
    * @param y2 - 顶点 2 Y 坐标
    * @param x3 - 顶点 3 X 坐标
    * @param y3 - 顶点 3 Y 坐标
-   * @param color - 边框颜色,默认白色,范围 0-1
-   * @param thickness - 线宽(像素),默认 1
+   * @param color - 边框颜色，默认白色，范围 0-1
+   * @param thickness - 线宽(像素)，默认 1
    */
   drawTriangle (x1: number, y1: number, x2: number, y2: number, x3: number, y3: number, color: Color = Color.WHITE, thickness: number = 1.0): void {
     this.setTriangleShape(x1, y1, x2, y2, x3, y3);
@@ -377,8 +371,8 @@ export class Graphics {
    * @param y - 矩形左下角 Y 坐标
    * @param width - 矩形宽度
    * @param height - 矩形高度
-   * @param color - 边框颜色,默认白色,范围 0-1
-   * @param thickness - 线宽(像素),默认 1
+   * @param color - 边框颜色，默认白色，范围 0-1
+   * @param thickness - 线宽(像素)，默认 1
    */
   drawRectangle (x: number, y: number, width: number, height: number, color: Color = Color.WHITE, thickness: number = 1.0): void {
     this.setRectangleShape(x, y, width, height, 0);
@@ -391,8 +385,8 @@ export class Graphics {
    * @param cx - 圆心 X 坐标
    * @param cy - 圆心 Y 坐标
    * @param radius - 半径(像素)
-   * @param color - 边框颜色,默认白色,范围 0-1
-   * @param thickness - 线宽(像素),默认 1
+   * @param color - 边框颜色，默认白色，范围 0-1
+   * @param thickness - 线宽(像素)，默认 1
    */
   drawCircle (cx: number, cy: number, radius: number, color: Color = Color.WHITE, thickness: number = 1.0): void {
     this.setCircleShape(cx, cy, radius);
@@ -401,14 +395,14 @@ export class Graphics {
   }
 
   /**
-   * 绘制填充三角形(实心)
+   * 绘制填充三角形（实心）
    * @param x1 - 顶点 1 X 坐标
    * @param y1 - 顶点 1 Y 坐标
    * @param x2 - 顶点 2 X 坐标
    * @param y2 - 顶点 2 Y 坐标
    * @param x3 - 顶点 3 X 坐标
    * @param y3 - 顶点 3 Y 坐标
-   * @param color - 填充颜色,默认白色,范围 0-1
+   * @param color - 填充颜色，默认白色，范围 0-1
    */
   fillTriangle (x1: number, y1: number, x2: number, y2: number, x3: number, y3: number, color: Color = Color.WHITE): void {
     this.setTriangleShape(x1, y1, x2, y2, x3, y3);
@@ -422,11 +416,11 @@ export class Graphics {
    * @param y - 矩形左下角 Y 坐标
    * @param width - 矩形宽度
    * @param height - 矩形高度
-   * @param color - 填充颜色,默认白色,范围 0-1
+   * @param color - 填充颜色，默认白色，范围 0-1
    *
-   * 直接用 pushQuad(2 三角形分割)而不是 setRectangleShape + buildShape,
-   * 后者走 Rectangle.triangulate 的中心 fan,中心顶点连到 4 角的 4 条对角内边
-   * 在 MSAA / 亚像素精度下会暴露 1-2 像素缝隙,叠层时底层颜色透出形成可见对角痕迹
+   * 直接用 pushQuad（2 三角形分割）而不是 setRectangleShape + buildShape，
+   * 后者走 Rectangle.triangulate 的中心 fan，中心顶点连到 4 角的 4 条对角内边
+   * 在 MSAA / 亚像素精度下会暴露 1-2 像素缝隙，叠层时底层颜色透出形成可见对角痕迹
    */
   fillRectangle (x: number, y: number, width: number, height: number, color: Color = Color.WHITE): void {
     this.ensureBatch('colored');
@@ -434,11 +428,11 @@ export class Graphics {
   }
 
   /**
-   * 绘制填充圆形(实心)
+   * 绘制填充圆形（实心）
    * @param cx - 圆心 X 坐标
    * @param cy - 圆心 Y 坐标
    * @param radius - 半径(像素)
-   * @param color - 填充颜色,默认白色,范围 0-1
+   * @param color - 填充颜色，默认白色，范围 0-1
    */
   fillCircle (cx: number, cy: number, radius: number, color: Color = Color.WHITE): void {
     this.setCircleShape(cx, cy, radius);
@@ -447,14 +441,14 @@ export class Graphics {
   }
 
   /**
-   * 绘制纹理矩形(本地坐标,Y 向上,(x, y) 为左下角)
+   * 绘制纹理矩形（本地坐标，Y 向上，(x, y) 为左下角）
    * @param x - 矩形左下角 X 坐标
    * @param y - 矩形左下角 Y 坐标
    * @param width - 矩形宽度
    * @param height - 矩形高度
-   * @param texture - 要采样的纹理。同纹理连续绘制会合批,纹理切换会自动 flush 当前批次
-   * @param region - 纹理 UV 子矩形,默认全图。Y 向上,(u0,v0) 为左下角 UV
-   * @param color - 乘色,默认白色
+   * @param texture - 要采样的纹理。同纹理连续绘制会合批，纹理切换会自动 flush 当前批次
+   * @param region - 纹理 UV 子矩形，默认全图。Y 向上，(u0,v0) 为左下角 UV
+   * @param color - 乘色，默认白色
    */
   drawTexture (
     x: number, y: number, width: number, height: number,
@@ -467,20 +461,20 @@ export class Graphics {
   }
 
   /**
-   * 绘制文本(本地坐标,Y 向上,(x, y) 为文本左下角)。
+   * 绘制文本（本地坐标，Y 向上，(x, y) 为文本左下角）。
    *
-   * 文本走字符级 bitmap atlas(同一字体下每个字只渲染一次,任意文本组合复用 atlas);
-   * `color` 作为乘色与白色字形 alpha 相乘,任意颜色都不会污染 atlas。
+   * 文本走字符级 bitmap atlas（同一字体下每个字只渲染一次，任意文本组合复用 atlas）;
+   * `color` 作为乘色与白色字形 alpha 相乘，任意颜色都不会污染 atlas。
    *
-   * 字体参数全部展开,避免调用方每帧创建临时 style 对象触发 GC
+   * 字体参数全部展开，避免调用方每帧创建临时 style 对象触发 GC
    * @param x - 文本左下角 X 坐标
-   * @param y - 文本左下角 Y 坐标(对齐 baseline 上方 ascent 处)
-   * @param text - 要绘制的文本内容,空串直接 return
-   * @param fontSize - 字号(逻辑像素)
-   * @param color - 乘色,默认白色,范围 0-1
-   * @param fontFamily - 字体族,默认 `sans-serif`
-   * @param fontWeight - 字重,默认 `normal`,支持 `'bold'` 或数字
-   * @param fontStyle - 字形,默认 `normal`,支持 `'italic'`
+   * @param y - 文本左下角 Y 坐标（对齐 baseline 上方 ascent 处）
+   * @param text - 要绘制的文本内容，空串直接 return
+   * @param fontSize - 字号（逻辑像素）
+   * @param color - 乘色，默认白色，范围 0-1
+   * @param fontFamily - 字体族，默认 `sans-serif`
+   * @param fontWeight - 字重，默认 `normal`，支持 `'bold'` 或数字
+   * @param fontStyle - 字形，默认 `normal`，支持 `'italic'`
    */
   drawText (
     x: number, y: number,
@@ -501,15 +495,15 @@ export class Graphics {
     const lineHeight = atlas.lineHeight;
     let cursorX = x;
 
-    // ensureChar 可能往 atlas canvas 写新字并打 dirty 标;实际 upload 推迟到
-    // flushBatch 统一处理,这样一帧多次 drawText 共写同一 atlas 只 upload 一次
+    // ensureChar 可能往 atlas canvas 写新字并打 dirty 标；实际 upload 推迟到
+    // flushBatch 统一处理，这样一帧多次 drawText 共写同一 atlas 只 upload 一次
     for (let i = 0; i < text.length; i++) {
       const info = atlas.ensureChar(text[i]);
 
       if (!info) {
         continue;
       }
-      // atlas 像素坐标 → UV(纹理 flipY 后,canvas 顶 → v=1,canvas 底 → v=0)
+      // atlas 像素坐标 → UV（纹理 flipY 后，canvas 顶 → v=1，canvas 底 → v=0）
       const u0 = info.px / ATLAS_SIZE;
       const u1 = (info.px + info.pw) / ATLAS_SIZE;
       const v0 = 1 - (info.py + info.ph) / ATLAS_SIZE;
@@ -594,13 +588,13 @@ export class Graphics {
   }
 
   /**
-   * 推一个 quad(4 顶点 + 6 索引,2 个三角形)。
+   * 推一个 quad（4 顶点 + 6 索引，2 个三角形）。
    *
-   * - colored 批次:`region` 省略 → UV 写 0(shader 不读)
-   * - textured 批次:传 `region` → UV 按 region 写
+   * - colored 批次：`region` 省略 → UV 写 0（shader 不读）
+   * - textured 批次：传 `region` → UV 按 region 写
    *
-   * 之所以用 2 三角形而不是中心 fan(像 Rectangle.triangulate),是因为 fan 的中心顶点连到 4 角的对角内边
-   * 在像素级亚像素精度下会暴露 1-2 像素缝隙,叠层时底层颜色透出形成可见的对角痕迹
+   * 之所以用 2 三角形而不是中心 fan（像 Rectangle.triangulate），是因为 fan 的中心顶点连到 4 角的对角内边
+   * 在像素级亚像素精度下会暴露 1-2 像素缝隙，叠层时底层颜色透出形成可见的对角痕迹
    */
   private pushQuad (
     x: number, y: number, width: number, height: number,
@@ -609,7 +603,7 @@ export class Graphics {
   ): void {
     const vertexOffset = this.vertices.length / 2;
 
-    // 4 顶点(本地坐标,左下/右下/左上/右上),应用当前变换写入 vertices
+    // 4 顶点（本地坐标，左下/右下/左上/右上），应用当前变换写入 vertices
     const corners: [number, number][] = [
       [x, y],
       [x + width, y],
